10/11/20       **Changes**       1. added register size to ConsumptionActivity       2. permission request in MainActivity       3. update compileSDK to 29       4. allow to android 10,11 versions to add directory and files to storage

+++ app/src/main/java/com/kp/meganet/meganetkp/ReadsActivity.java
@@ -574,7 +574,7 @@ public class ReadsActivity extends AppCompatActivity implements iReadMeterCallBa
         int[] intDataArr = new int[data_prm.length];
 
         for (int i = 0; i < data_prm.length; i++)
-            intDataArr[i] = (int) (data_prm[i] & 0xff);
+            intDataArr[i] = data_prm[i] & 0xff;
 
         // Rows.Add(Format(Int(tm.Hour), "##00") & ":" & Format(Int(tm.Minute), "##00") & ":" & Format(Int(tm.Second), "##00"), ID, System, Collector, Level)
         if (intDataArr.length == 13) {
